+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 - **Breaking**: 最小サポートをbeta42に引き上げ
+- **Breaking**: EditSectionObjectCaller、EditSectionLayerCallerに型引数を追加
 - filter: ObjectInfo.is_filter_objectを追加
+- generic: ReadSectionを追加
 
 ### デモプラグイン